+++ packages/connector-helper/src/wallet-connection.ts
@@ -1,0 +1,8 @@
+import type { BlockchainWallet } from "@rarible/sdk-wallet"
+import type { Blockchain } from "@rarible/api-client"
+
+export interface IWalletAndAddress {
+	wallet: BlockchainWallet
+	address: string
+	blockchain: Blockchain
+}