+++ packages/sdk/src/config/development.ts
@@ -3,7 +3,7 @@ import type { RaribleSdkConfig } from "./domain"
 export const developmentConfig: RaribleSdkConfig = {
 	basePath: "https://dev-api.rarible.org",
 	ethereumEnv: "dev-ethereum",
-	flowEnv: "dev",
+	flowEnv: "dev-testnet",
 	tezosNetwork: "dev",
 	polygonNetwork: "dev-polygon",
 }