+++ packages/sdk/src/index.ts
@@ -1,9 +1,7 @@
-import type { BlockchainWallet } from "@rarible/sdk-wallet"
 import type { ContractAddress } from "@rarible/types"
-import { toContractAddress } from "@rarible/types"
 import type { Maybe } from "@rarible/types/build/maybe"
-import { BlockchainGroup } from "@rarible/api-client"
-import type { WalletByBlockchain } from "@rarible/sdk-wallet/src"
+import { Blockchain, BlockchainGroup } from "@rarible/api-client"
+import type { BlockchainWallet, WalletByBlockchain } from "@rarible/sdk-wallet"
 import type { IApisSdk, IRaribleInternalSdk, IRaribleSdk, IRaribleSdkConfig, ISdkContext } from "./domain"
 import { LogsLevel } from "./domain"
 import { getSdkConfig } from "./config"
@@ -28,13 +26,18 @@ export function createRaribleSdk(
 ): IRaribleSdk {
 	const blockchainConfig = getSdkConfig(env)
 	const apis = createApisSdk(env, config?.apiClientParams)
+	const ethConfig = {
+		params: config?.apiClientParams,
+		logs: config?.logs ?? LogsLevel.TRACE,
+		ethereum: config?.ethereum,
+		polygon: config?.polygon,
+	}
 	const instance = createUnionSdk(
 		createEthereumSdk(
 			filterWallet(wallet, BlockchainGroup.ETHEREUM),
 			apis,
 			blockchainConfig.ethereumEnv,
-			config?.apiClientParams,
-			config?.logs ?? LogsLevel.TRACE
+			ethConfig
 		),
 		createFlowSdk(
 			filterWallet(wallet, BlockchainGroup.FLOW),
@@ -50,8 +53,7 @@ export function createRaribleSdk(
 			filterWallet(wallet, BlockchainGroup.ETHEREUM),
 			apis,
 			blockchainConfig.polygonNetwork,
-			config?.apiClientParams,
-			config?.logs ?? LogsLevel.TRACE
+			ethConfig
 		),
 	)
 
@@ -118,8 +120,7 @@ function filterWallet<T extends BlockchainGroup>(
 function createSell(sell: ISellInternal, apis: IApisSdk): ISell {
 	return async ({ itemId }) => {
 		const item = await apis.item.getItemById({ itemId })
-		const collectionId = toContractAddress(item.contract)
-		const response = await sell({ collectionId })
+		const response = await sell({ blockchain: item.blockchain })
 		return {
 			...response,
 			maxAmount: item.supply,
@@ -136,7 +137,8 @@ function createMintAndSell(mint: IMint, sell: ISellInternal): IMintAndSell {
 	return async request => {
 		const mintResponse = await mint(request)
 		const collectionId = getCollectionId(request)
-		const sellResponse = await sell({ collectionId })
+		const blockchain = getBlockchainCollectionId(collectionId)
+		const sellResponse = await sell({ blockchain })
 
 		const mintAction = mintResponse.submit
 			.around(
@@ -170,6 +172,14 @@ export function getCollectionId(req: HasCollectionId | HasCollection): ContractA
 		return req.collection.id
 	}
 	return req.collectionId
+}
+
+function getBlockchainCollectionId(contract: ContractAddress): Blockchain {
+	const [blockchain] = contract.split(":")
+	if (!(blockchain in Blockchain)) {
+		throw new Error(`Unrecognized blockchain in contract ${contract}`)
+	}
+	return blockchain as Blockchain
 }
 
 type MiddleMintType = {