+++ packages/sdk/src/sdk-blockchains/ethereum/balance.test.ts
@@ -1,7 +1,6 @@
 import { Web3Ethereum } from "@rarible/web3-ethereum"
 import { EthereumWallet } from "@rarible/sdk-wallet"
-import { toContractAddress, toUnionAddress } from "@rarible/types"
-import { awaitAll, deployTestErc20 } from "@rarible/ethereum-sdk-test-common"
+import { toContractAddress, toCurrencyId, toUnionAddress, ZERO_ADDRESS } from "@rarible/types"
 import type { AssetType } from "@rarible/api-client"
 import { Blockchain } from "@rarible/api-client"
 import BigNumber from "bignumber.js"
@@ -20,11 +19,7 @@ describe("get balance", () => {
 		web3: web31,
 		from: wallet1.getAddressString(),
 	})
-	const sdk = createRaribleSdk(new EthereumWallet(ethereum), "e2e", { logs: LogsLevel.DISABLED })
-
-	const it = awaitAll({
-		testErc20: deployTestErc20(web31, "Test1", "TST1"),
-	})
+	const sdk = createRaribleSdk(new EthereumWallet(ethereum), "development", { logs: LogsLevel.DISABLED })
 
 	test("get ETH balance with wallet", async () => {
 		const walletAddress = toUnionAddress("ETHEREUM:0xa14FC5C72222FAce8A1BcFb416aE2571fA1a7a91")
@@ -35,7 +30,7 @@ describe("get balance", () => {
 	})
 
 	test("get ETH balance without wallet", async () => {
-		const sdk = createRaribleSdk(undefined, "e2e", { logs: LogsLevel.DISABLED })
+		const sdk = createRaribleSdk(undefined, "development", { logs: LogsLevel.DISABLED })
 		const walletAddress = toUnionAddress("ETHEREUM:0xa14FC5C72222FAce8A1BcFb416aE2571fA1a7a91")
 		const balance = await sdk.balances.getBalance(walletAddress, {
 			"@type": "ETH",
@@ -43,28 +38,31 @@ describe("get balance", () => {
 		expect(balance.toString()).toEqual("1.9355")
 	})
 
+	test("get ETH balance without wallet with CurrencyId", async () => {
+		const sdk = createRaribleSdk(undefined, "development", { logs: LogsLevel.DISABLED })
+		const walletAddress = toUnionAddress("ETHEREUM:0xa14FC5C72222FAce8A1BcFb416aE2571fA1a7a91")
+		const currency = toCurrencyId(`ETHEREUM:${ZERO_ADDRESS}`)
+		const balance = await sdk.balances.getBalance(walletAddress, currency)
+		expect(balance.toString()).toEqual("1.9355")
+	})
+
 	test("get balance erc-20", async () => {
-		expect.assertions(1)
-		const senderRaw = wallet1.getAddressString()
-		const sender = toUnionAddress(`ETHEREUM:${senderRaw}`)
-		await it.testErc20.methods.mint(senderRaw, 1).send({
-			from: senderRaw,
-			gas: 200000,
-		})
+		const sender = toUnionAddress("ETHEREUM:0xa14FC5C72222FAce8A1BcFb416aE2571fA1a7a91")
 
-		const contract = toContractAddress(`ETHEREUM:${it.testErc20.options.address}`)
-		const nextBalance = "0.000000000000000001"
-		const balance = await retry(5, 1000, async () => {
-			const balance = await sdk.balances.getBalance(sender, {
-				"@type": "ERC20",
-				contract,
-			})
-			if (balance.toString() !== nextBalance) {
-				throw new Error("Unequal balances")
-			}
-			return balance
+		const contract = toContractAddress("ETHEREUM:0x55eB2809896aB7414706AaCDde63e3BBb26e0BC6")
+		const nextBalance = "0.00035"
+		const balance = await sdk.balances.getBalance(sender, {
+			"@type": "ERC20",
+			contract,
 		})
+		expect(balance.toString()).toEqual(nextBalance)
+	})
 
+	test("get balance erc-20 with CurrencyId", async () => {
+		const sender = toUnionAddress("ETHEREUM:0xa14FC5C72222FAce8A1BcFb416aE2571fA1a7a91")
+		const contract = toCurrencyId("ETHEREUM:0x55eB2809896aB7414706AaCDde63e3BBb26e0BC6")
+		const nextBalance = "0.00035"
+		const balance = await sdk.balances.getBalance(sender, contract)
 		expect(balance.toString()).toEqual(nextBalance)
 	})
 
@@ -72,22 +70,22 @@ describe("get balance", () => {
 		const senderRaw = wallet1.getAddressString()
 		const wethE2eAssetType: AssetType = {
 			"@type": "ERC20",
-			contract: convertEthereumContractAddress("0xc6f33b62a94939e52e1b074c4ac1a801b869fdb2", Blockchain.ETHEREUM),
+			contract: convertEthereumContractAddress("0x55eB2809896aB7414706AaCDde63e3BBb26e0BC6", Blockchain.ETHEREUM),
 		}
 		const sender = convertEthereumToUnionAddress(senderRaw, Blockchain.ETHEREUM)
 		const initWethBalance = await sdk.balances.getBalance(sender, wethE2eAssetType)
 		const convertTx = await sdk.balances.convert({
 			blockchain: Blockchain.ETHEREUM,
 			isWrap: true,
-			value: "0.00000000000035",
+			value: "0.00035",
 		})
 		await convertTx.wait()
 
-		await retry(5, 2000, async () => {
+		await retry(10, 2000, async () => {
 			const finishWethBalance = await sdk.balances.getBalance(sender, wethE2eAssetType)
 
 			expect(finishWethBalance.toString()).toBe(
-				new BigNumber(initWethBalance).plus("0.00000000000035").toString()
+				new BigNumber(initWethBalance).plus("0.00035").toString()
 			)
 		})
 	})
@@ -96,7 +94,7 @@ describe("get balance", () => {
 		const senderRaw = wallet1.getAddressString()
 		const wethE2eAssetType: AssetType = {
 			"@type": "ERC20",
-			contract: convertEthereumContractAddress("0xc6f33b62a94939e52e1b074c4ac1a801b869fdb2", Blockchain.ETHEREUM),
+			contract: convertEthereumContractAddress("0x55eB2809896aB7414706AaCDde63e3BBb26e0BC6", Blockchain.ETHEREUM),
 		}
 		const sender = convertEthereumToUnionAddress(senderRaw, Blockchain.ETHEREUM)
 		const balanceWithoutWeth = await sdk.balances.getBalance(sender, wethE2eAssetType)
@@ -142,6 +140,13 @@ describe("get polygon balance", () => {
 			"@type": "ETH",
 			blockchain: Blockchain.POLYGON,
 		})
+		expect(balance.toString()).toEqual("0.009145")
+	})
+
+	test("get Matic balance with CurrencyId", async () => {
+		const walletAddress = toUnionAddress("ETHEREUM:0xc8f35463Ea36aEE234fe7EFB86373A78BF37e2A1")
+		const currency = toCurrencyId(`POLYGON:${ZERO_ADDRESS}`)
+		const balance = await sdk.balances.getBalance(walletAddress, currency)
 		expect(balance.toString()).toEqual("0.009145")
 	})