+++ packages/sdk/src/sdk-blockchains/ethereum/common/index.ts
@@ -15,13 +15,11 @@ import type { UnionPart } from "packages/sdk/src/types/order/common"
 import type { ContractAddress } from "@rarible/types/build/contract-address"
 import type { EthereumNetwork } from "@rarible/protocol-ethereum-sdk/build/types"
 import { toBn } from "@rarible/utils/build/bn"
-import type {
-	AssetType as EthereumAssetType,
-	Part,
-} from "@rarible/ethereum-api-client"
+import type { AssetType as EthereumAssetType, Part } from "@rarible/ethereum-api-client"
 import type { Order } from "@rarible/ethereum-api-client/build/models"
 import type { EthereumTransaction } from "@rarible/ethereum-provider"
-import type { CurrencyType, RequestCurrency } from "../../../common/domain"
+import type { CurrencyType } from "../../../common/domain"
+import type { RequestCurrencyAssetType } from "../../../common/domain"
 
 export type EVMBlockchain = Blockchain.ETHEREUM | Blockchain.POLYGON
 export const EVMBlockchains: EVMBlockchain[] = [
@@ -31,7 +29,7 @@ export const EVMBlockchains: EVMBlockchain[] = [
 
 export type CreateEthereumCollectionResponse = { tx: EthereumTransaction, address: Address }
 
-export function getEthTakeAssetType(currency: RequestCurrency) {
+export function getEthTakeAssetType(currency: RequestCurrencyAssetType) {
 	switch (currency["@type"]) {
 		case "ERC20":
 			return {