+++ packages/sdk/src/sdk-blockchains/tezos/sell.ts
@@ -6,8 +6,6 @@ import { Action } from "@rarible/action"
 import type { TezosProvider, FTAssetType, XTZAssetType } from "@rarible/tezos-sdk"
 import BigNumber from "bignumber.js"
 import type { OrderForm } from "@rarible/tezos-sdk/dist/order"
-import { Blockchain } from "@rarible/api-client"
-import type { RequestCurrency } from "../../common/domain"
 import { OriginFeeSupport, PayoutsSupport } from "../../types/order/fill/domain"
 import type * as OrderCommon from "../../types/order/common"
 import { retry } from "../../common/retry"
@@ -16,6 +14,9 @@ import type {
 	PrepareOrderUpdateRequest,
 	PrepareOrderUpdateResponse,
 } from "../../types/order/common"
+import type { RequestCurrencyAssetType } from "../../common/domain"
+import { getCurrencyAssetType } from "../../common/get-currency-asset-type"
+import type { PrepareSellInternalResponse } from "../../types/order/sell/domain"
 import type { TezosOrder } from "./domain"
 import type { ITezosAPI, MaybeProvider } from "./common"
 import {
@@ -37,7 +38,7 @@ export class TezosSell {
 		this.update = this.update.bind(this)
 	}
 
-	parseTakeAssetType(type: RequestCurrency): XTZAssetType | FTAssetType {
+	parseTakeAssetType(type: RequestCurrencyAssetType): XTZAssetType | FTAssetType {
 		switch (type["@type"]) {
 			case "XTZ":
 				return {
@@ -55,28 +56,22 @@ export class TezosSell {
 		}
 	}
 
-	async sell(
-		prepareRequest: OrderCommon.PrepareOrderInternalRequest
-	): Promise<OrderCommon.PrepareOrderInternalResponse> {
-		const [domain, contract] = prepareRequest.collectionId.split(":")
-		if (domain !== Blockchain.TEZOS) {
-			throw new Error("Not an tezos item")
-		}
-		const itemCollection = await this.apis.collection.getNftCollectionById({
-			collection: contract,
-		})
-
+	async sell(): Promise<PrepareSellInternalResponse> {
 		const submit = Action.create({
 			id: "send-tx" as const,
 			run: async (request: OrderCommon.OrderInternalRequest) => {
 				const provider = getRequiredProvider(this.provider)
 				const makerPublicKey = await getMakerPublicKey(provider)
-				const { itemId } = getTezosItemData(request.itemId)
+				const { itemId, contract } = getTezosItemData(request.itemId)
 
 				const item = await retry(90, 1000, async () => {
 				   return this.apis.item.getNftItemById({ itemId })
 				})
+				const requestCurrency = getCurrencyAssetType(request.currency)
 
+				const itemCollection = await this.apis.collection.getNftCollectionById({
+					collection: contract,
+				})
 				const tezosRequest: TezosSellRequest = {
 					maker: pk_to_pkh(makerPublicKey),
 					maker_edpk: makerPublicKey,
@@ -85,7 +80,7 @@ export class TezosSell {
 						contract: item.contract,
 						token_id: new BigNumber(item.tokenId),
 					},
-					take_asset_type: this.parseTakeAssetType(request.currency),
+					take_asset_type: this.parseTakeAssetType(requestCurrency),
 					amount: new BigNumber(request.amount),
 					price: new BigNumber(request.price),
 					payouts: await getPayouts(provider, request.payouts),
@@ -101,7 +96,6 @@ export class TezosSell {
 		})
 
 		return {
-			multiple: itemCollection.type === "MT",
 			originFeeSupport: OriginFeeSupport.FULL,
 			payoutsSupport: PayoutsSupport.MULTIPLE,
 			supportedCurrencies: getSupportedCurrencies(),