+++ scripts/build.sh
@@ -13,3 +13,4 @@ yarn run build-connector-torus
 yarn run build-connector-walletlink
 yarn run build-connector-walletconnect
 yarn run build-connector-phantom
+yarn run build-connector-helper